RAC-153 refactor : 대학생 멘토링 완료 변경시 급여 증가, 대학원생 수락시 기대 날짜 변경

+++ src/main/java/com/postgraduate/domain/mentoring/application/dto/WaitingSeniorMentoringInfo.java
@@ -12,5 +12,5 @@ public class WaitingSeniorMentoringInfo {
     private String profile;
     private String nickname;
     private int term;
-    private int remainTime;
+    private long remainTime;
 }
